JobsTableViewController TableView Extension
